Begin breaking up arch.hpp
