+++ kernel/arch/i386/gdt.cpp
@@ -10,6 +10,7 @@
  * @copyright Copyright the Xyris Contributors (c) 2019
  *
  */
+#include <arch/i386/gdt.hpp>
 #include <arch/arch.hpp>
 #include <lib/string.hpp>
 #include <lib/stdio.hpp>
