+++ kernel/arch/i386/i386.cpp
@@ -1,0 +1,11 @@
+/**
+ * @file i386.cpp
+ * @author Keeton Feavel (keetonfeavel@cedarville.edu)
+ * @brief 
+ * @version 0.1
+ * @date 2021-08-24
+ * 
+ * @copyright Copyright the Xyris Contributors (c) 2021
+ * 
+ */
+#include <arch/i386/i386.hpp>
