+++ kernel/arch/i386/ports.cpp
@@ -8,7 +8,7 @@
  * @copyright Copyright the Xyris Contributors (c) 2020
  *
  */
-#include <arch/arch.hpp>
+#include <arch/i386/ports.hpp>
 
 uint8_t readByte(uint16_t port) {
     uint8_t result;
