+++ kernel/arch/i386/ports.hpp
@@ -15,7 +15,6 @@
  *
  */
 #pragma once
-
 #include <stdint.h>
 
 /**
