+++ kernel/dev/serial/rs232.cpp
@@ -13,7 +13,8 @@
  */
 
 #include <stdarg.h>
-#include <arch/arch.hpp>
+#include <arch/i386/ports.hpp>
+#include <arch/i386/isr.hpp>
 #include <dev/serial/rs232.hpp>
 #include <mem/heap.hpp>
 #include <lib/stdio.hpp>
